OpenCL: If OpenCL 1.2 is not requested, use the default
OpenCL 2.0 is not the best in terms of portability due to
the lack of platforms that support its extensive feature
set.  OpenCL 3.0 could be a better default as it's more
flexible in terms of the features, but I don't so far
see the need to force any standard by default.