OpenCL: Workaround for Intel OpenCL driver for CPU
Intel CPU OpenCL driver for some reason declares the cl_amd_media_ops
built-ins and includes the 'overloadable' attribute in the declarations.
Since these definitions do not use the 'overloadable' (it's not standard
OpenCL C), the definitions are not compatible with them, causing a build
error due to a mismatching declaration.