OpenCL: fix reliance on 32b wraparound
A couple of spots in code perform an uint mad24 with
arguments that might have negative values, leading to
a negative end result.  Since the mad24 called is unsigned,
the offset will be huge and will cause the kernel to
crash when executed on a machine with 64b pointer arithmetics.
This caused crashes both on Intel CPU driver and PoCL CPU
driver.

This likely has worked on machines with 32b offsetting as
the wraparound brings the value to the (intended?) negative
offset.

Fix here by casting the mad24 result to a 32b int. I'm a bit
unsure if this is a correct fix here, so the original
author should review it carefully before pulled in.