PoCL - OpenCL portability fixes (#1505)
* OpenCL: Fall back to any device if GPUs are not found

This allows attempting to run on CPUs and other type devices.

* OpenCL: dynamic storage for CL_DEVICE_EXTENSIONS

PoCL and Intel GPU OpenCL drivers support a lot of extensions.
Now any size extensions list can be fetched.

* OpenCL: If OpenCL 1.2 is not requested, use the default

OpenCL 2.0 is not the best in terms of portability due to
the lack of platforms that support its extensive feature
set.  OpenCL 3.0 could be a better default as it's more
flexible in terms of the features, but I don't so far
see the need to force any standard by default.

* OpenCL: Workaround for Intel OpenCL driver for CPU

Intel CPU OpenCL driver for some reason declares the cl_amd_media_ops
built-ins and includes the 'overloadable' attribute in the declarations.
Since these definitions do not use the 'overloadable' (it's not standard
OpenCL C), the definitions are not compatible with them, causing a build
error due to a mismatching declaration.

* OpenCL: use vload/vstore for unaligned vector loads/stores

This makes the loads and stores defined behavior.

Fixes #1484.

* OpenCL: fix reliance on 32b wraparound

A couple of spots in code perform an uint mad24 with
arguments that might have negative values, leading to
a negative end result.  Since the mad24 called is unsigned,
the offset will be huge and will cause the kernel to
crash when executed on a machine with 64b pointer arithmetics.
This caused crashes both on Intel CPU driver and PoCL CPU
driver.

This likely has worked on machines with 32b offsetting as
the wraparound brings the value to the (intended?) negative
offset.

Fix here by casting the mad24 result to a 32b int. I'm a bit
unsure if this is a correct fix here, so the original
author should review it carefully before pulled in.

* OpenCL: Round to even integer before clamping

Some of the CTS tests expect this behavior.

After this amd_pack() behavior matches the AMD GPU SDK behavior with
that intrinsic.

---------

Co-authored-by: Kiriti Gowda <kiritigowda@gmail.com>